confirm before deleting branch

+++ src/git/merge.rs
@@ -1,7 +1,7 @@
 use crate::arg::save;
 // use crate::error::unimplemented;
 use crate::git::{branch, checkout, commit::commit, pull, status::is_status_clean};
-use crate::misc::warning;
+use crate::misc::{confirm, warning};
 use crate::proc;
 use crate::Opt;
 use std::collections::HashMap;
@@ -39,7 +39,10 @@ pub fn merge_not_master<S: ::std::hash::BuildHasher + Default>(args: &HashMap<St
         pull::pull(&args_c["remote"].value, &args_c["branch"].value, false);
         let command = format!("git merge {} --no-ff", branch);
         proc::execute(&command);
-        branch::delete_branch(&branch);
+        let command = format!("Delete branch {}", branch);
+        if confirm(&command) {
+            branch::delete_branch(&branch);
+        }
         if Path::new("./.config.toml").exists() {
             save(&args_c);
         }
